# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReadinessStatusChecks = {
  database: string;
};

export interface ReadinessStatus {
  status: string;
  checks: ReadinessStatusChecks;
}

export type GamePlatform = typeof GamePlatform[keyof typeof GamePlatform];


export const GamePlatform = {
  PS5_ONLY: 'PS5_ONLY',
  PS4_AND_PS5: 'PS4_AND_PS5',
  PS4_ONLY: 'PS4_ONLY',
} as const;

export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;

export interface Game {
  id: string;
  title: string;
  titleNormalized: string;
  coverUrl?: string | null;
  platform: GamePlatform;
  status: GameStatus;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | null;
}

export type GameListItem = Game & {
  accountCount: number;
};

export interface GameListResponse {
  games: GameListItem[];
}

export interface GameResponse {
  game: Game;
}

export interface CreateGameRequest {
  /**
     * Game title. Must be 1-120 characters and cannot be whitespace-only.
     * @minLength 1
     * @maxLength 120
     * @pattern ^(?!\s*$).+$
     */
  title: string;
  coverUrl?: string;
  platform: GamePlatform;
  status?: GameStatus;
}

export interface UpdateGameRequest {
  /**
     * Game title. Must be 1-120 characters and cannot be whitespace-only.
     * @minLength 1
     * @maxLength 120
     * @pattern ^(?!\s*$).+$
     */
  title?: string;
  coverUrl?: string | null;
  platform?: GamePlatform;
  status?: GameStatus;
}

export interface OkResponse {
  ok: boolean;
}

export type ErrorResponseIssuesItem = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  issues?: ErrorResponseIssuesItem[];
}

export type AccountStatus = typeof AccountStatus[keyof typeof AccountStatus];


export const AccountStatus = {
  AVAILABLE: 'AVAILABLE',
  PARTIALLY_SOLD: 'PARTIALLY_SOLD',
  SOLD: 'SOLD',
  INACTIVE: 'INACTIVE',
} as const;

export type CapacityKind = typeof CapacityKind[keyof typeof CapacityKind];


export const CapacityKind = {
  Z2_PS5: 'Z2_PS5',
  Z2_PS4: 'Z2_PS4',
  Z3_SHARED_PS5_PS4: 'Z3_SHARED_PS5_PS4',
} as const;

export interface AccountListItem {
  id: string;
  gameId: string;
  accountCode: string;
  accountNumberPrefix: string;
  accountNumberSeq: number;
  displayNumber: string;
  onlineId: string | null;
  birthDate: string | null;
  status: AccountStatus;
  createdAt: string;
  updatedAt: string;
}

export interface AccountCapacity {
  id: string;
  accountId: string;
  capacityKind: CapacityKind;
  instanceNo: number;
  displayLabel: string;
  isFinished: boolean;
  finishedAt: string | null;
}

export type AccountDetail = AccountListItem & {
  capacities: AccountCapacity[];
};

export interface AccountListResponse {
  accounts: AccountListItem[];
}

export interface AccountDetailResponse {
  account: AccountDetail;
}

export interface AccountCapacitiesResponse {
  capacities: AccountCapacity[];
  status: AccountStatus;
}

export interface CreateAccountRequest {
  psnEmail: string;
  /** @minLength 1 */
  psnPassword: string;
  /** @minLength 1 */
  emailPassword: string;
  /** @minLength 1 */
  onlineId: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  birthDate: string;
  familyManagementEmail: string;
  /**
     * @minItems 1
     * @items.minLength 1
     */
  backupCodes: string[];
  confirmed?: boolean;
}

export type DuplicateWarningResponseCode = typeof DuplicateWarningResponseCode[keyof typeof DuplicateWarningResponseCode];


export const DuplicateWarningResponseCode = {
  DUPLICATE_WARNING: 'DUPLICATE_WARNING',
} as const;

export type DuplicateFieldName = typeof DuplicateFieldName[keyof typeof DuplicateFieldName];


export const DuplicateFieldName = {
  psnEmail: 'psnEmail',
  familyManagementEmail: 'familyManagementEmail',
  onlineId: 'onlineId',
} as const;

export type DuplicateWarningResponseDetail = {
  duplicateFields: DuplicateFieldName[];
};

export interface DuplicateWarningResponse {
  error: string;
  code: DuplicateWarningResponseCode;
  detail: DuplicateWarningResponseDetail;
}

export type StandardApiErrorIssuesItem = { [key: string]: unknown };

export interface StandardApiError {
  /** Persian user-safe message */
  error: string;
  /** Machine-readable error code */
  code: string;
  issues?: StandardApiErrorIssuesItem[];
}