# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games/{gameId}/accounts:
    get:
      operationId: listAccounts
      tags:
        - accounts
      summary: List all accounts for a game
      description: Returns non-secret Account summaries for a single Game.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Accounts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
        '400':
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
    post:
      operationId: createAccount
      tags:
        - accounts
      summary: Create an account for a game
      description: >-
        Creates a new Account, its Capacity template, and its encrypted Backup Codes. The Game ID is taken from the path
        only. Duplicate values may be accepted by sending the same request again with `confirmed: true`.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetailResponse'
        '400':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '403':
          description: Account operations disabled in the mounted runtime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '409':
          description: Duplicate warning, inactive game, or identifier conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicateWarningResponse'
        '500':
          description: Unexpected or encryption configuration failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
  /accounts/{accountId}:
    get:
      operationId: getAccount
      tags:
        - accounts
      summary: Get a single account
      description: Returns safe Account metadata without any Secret.
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetailResponse'
        '400':
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
  /accounts/{accountId}/capacities:
    get:
      operationId: getAccountCapacities
      tags:
        - accounts
      summary: Get account capacities
      description: Returns Capacity rows and derived Account status information required by the Account workspace.
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account capacities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountCapacitiesResponse'
        '400':
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardApiError'
components:
  schemas:
...
    AccountStatus:
      type: string
      enum:
        - AVAILABLE
        - PARTIALLY_SOLD
        - SOLD
        - INACTIVE
    CapacityKind:
      type: string
      enum:
        - Z2_PS5
        - Z2_PS4
        - Z3_SHARED_PS5_PS4
    AccountListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        accountCode:
          type: string
        accountNumberPrefix:
          type: string
        accountNumberSeq:
          type: integer
        displayNumber:
          type: string
        onlineId:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - gameId
        - accountCode
        - accountNumberPrefix
        - accountNumberSeq
        - displayNumber
        - onlineId
        - birthDate
        - status
        - createdAt
        - updatedAt
    AccountDetail:
      allOf:
        - $ref: '#/components/schemas/AccountListItem'
        - type: object
          properties:
            capacities:
              type: array
              items:
                $ref: '#/components/schemas/AccountCapacity'
          required:
            - capacities
    AccountCapacity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        capacityKind:
          $ref: '#/components/schemas/CapacityKind'
        instanceNo:
          type: integer
        displayLabel:
          type: string
        isFinished:
          type: boolean
        finishedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - accountId
        - capacityKind
        - instanceNo
        - displayLabel
        - isFinished
        - finishedAt
    AccountListResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountListItem'
      required:
        - accounts
    AccountDetailResponse:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/AccountDetail'
      required:
        - account
    AccountCapacitiesResponse:
      type: object
      properties:
        capacities:
          type: array
          items:
            $ref: '#/components/schemas/AccountCapacity'
        status:
          $ref: '#/components/schemas/AccountStatus'
      required:
        - capacities
        - status
    CreateAccountRequest:
      type: object
      properties:
        psnEmail:
          type: string
          format: email
        psnPassword:
          type: string
          minLength: 1
        emailPassword:
          type: string
          minLength: 1
        onlineId:
          type: string
          minLength: 1
        birthDate:
          type: string
          pattern: ^\d{4}-\d{2}-\d{2}$
        familyManagementEmail:
          type: string
          format: email
        backupCodes:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
        confirmed:
          type: boolean
          default: false
      required:
        - psnEmail
        - psnPassword
        - emailPassword
        - onlineId
        - birthDate
        - familyManagementEmail
        - backupCodes
    DuplicateWarningResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum:
            - DUPLICATE_WARNING
        detail:
          type: object
          properties:
            duplicateFields:
              type: array
              items:
                $ref: '#/components/schemas/DuplicateFieldName'
          required:
            - duplicateFields
      required:
        - error
        - code
        - detail
    StandardApiError:
      type: object
      properties:
        error:
          type: string
          description: Persian user-safe message
        code:
          type: string
          description: Machine-readable error code
        issues:
          type: array
          items:
            type: object
      required:
        - error
        - code
    DuplicateFieldName:
      type: string
      enum:
        - psnEmail
        - familyManagementEmail
        - onlineId