# lib/api-zod/src/generated/types/createAccountRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAccountRequest {
  psnEmail: string;
  /** @minLength 1 */
  psnPassword: string;
  /** @minLength 1 */
  emailPassword: string;
  /** @minLength 1 */
  onlineId: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  birthDate: string;
  familyManagementEmail: string;
  /**
     * @minItems 1
     * @items.minLength 1
     */
  backupCodes: string[];
  confirmed?: boolean;
}