# lib/api-zod/src/generated/types/duplicateFieldName.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type DuplicateFieldName = typeof DuplicateFieldName[keyof typeof DuplicateFieldName];


export const DuplicateFieldName = {
  psnEmail: 'psnEmail',
  familyManagementEmail: 'familyManagementEmail',
  onlineId: 'onlineId',
} as const;