# lib/api-zod/src/generated/types/duplicateWarningResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DuplicateWarningResponseCode } from './duplicateWarningResponseCode';
import type { DuplicateWarningResponseDetail } from './duplicateWarningResponseDetail';

export interface DuplicateWarningResponse {
  error: string;
  code: DuplicateWarningResponseCode;
  detail: DuplicateWarningResponseDetail;
}